New end of inflation (#14)
* Closes #8 

I did some thinking related to the two options you suggested.

In this regard since we have the analysis now using \ddot a=0 condition, we can replace figs 1 and 2. Fig. 3 we leave as is because it is just an illustrative figure.

So I have worked on completing the reply to the referee report and the new version is attached. The text in red in item 4) is new and so please read and comment.

I have also added some text in the revised version of the paper related to item 4 and item 6.